분석 1트 완료
